disk: Ensure we use the explicit path to resize2fs
On some configurations, /sbin and /usr/sbin may not even be in the $PATH,
meaning that the configure routine would fail if this is where resize2fs
lives. Secondly, we must ensure that disks now uses this explicit path.

This ensures correct functionality during non privileged testing, and
consistency with installed behaviour.

Signed-off-by: Ikey Doherty <michael.i.doherty@intel.com>