Fix the user-data passwd field
The contents of the passwd field need to be quoted before being
passed to useradd.  Password hashes often contain $ symbols and
without the quotes these symbols are interpeted as shell variables.
The result being that the hash you specify in your yaml bears
little resemblance to the hash that ends up in /etc/shadow and
your user cannot log in.

Signed-off-by: Mark Ryan <mark.d.ryan@intel.com>