Small build fixes (#19)
* Make -Werror a compile time option

Signed-off-by: Ikey Doherty <michael.i.doherty@intel.com>

* Ensure test suite builds with -Wl,--as-needed

The linking order is highly important to libtool. Previously we would link
the COMMON_LDADD libraries to the binaries, however no symbols were used so
binutils linker wouldn't link the libraries involved. Next we would link
libtest.la, a static convenience library, and fail at dynamic linking due
to the symbols only being used in this later-added libraries.

The solution is to first link the convenience library, and then link to the
shared libraries, in order for -Wl,--as-needed behaviour to function correctly
and only link the absolutely required libraries.

Signed-off-by: Ikey Doherty <michael.i.doherty@intel.com>