Remove forced dependency ordering on systemd-networkd.
These lines forced systemd-networkd to wait until `ucd` had
completed.

In turn, `ucd` requires the network to be active in order
to install packages. It waits for the network to become
active, even.

Surely, this is untenable. Ucd will have to accept any
network state and may not prevent it from starting as
early as possible, which would be counter to any fast
boot initiative.