Add equinix provisioning template
To provision on equinix we can grab the `keys` file from their
service. The address I entered in the template is tentative - it
needs to be checked that metadata.platformequinix.com resolves to that
address in their internal network as well that it's on a static IP
address and not some RR load balancer service - ideally.